Starting logical part
Reads inputs from user, select the piece to move, verify if this piece exists in the board in the introduced position and put an empty character on initial position.
Next steps are select a new position to put the piece, verify if the piece can move to this position and then move it to new position.